rename src -> components
