+++ components/CatalogFavorite/CatalogFavorite.js
