+++ components/SameWines/index.js
@@ -1,2 +1,3 @@
 import SameWines from './SameWines'
+
 export default SameWines
