+++ pages/profile/Profile.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import Link from 'next/link'
 import { useRecoilValue } from 'recoil'
-import Header from '../../src/Header'
-import Search from '../../src/Search'
+import Header from '../../components/Header'
+import Search from '../../components/Search'
 import { userState } from '../../store/GlobalRecoilWrapper/store'
 import useLocalStorage from '../../utils/useLocalStorage'
